# 🚀 use atomic.Value instead of sync.Mutex