**Please provide enough information so that others can review your pull request:**
To improve the performance and precision
<!-- You can skip this if you're fixing a typo or adding an app to the Showcase. -->

**Explain the *details* for making this change. What existing problem does the pull request solve?**
`time.Sleep(500 * time.Millisecond)` will impact precision when user set the time format which accurate to milliseconds.
<!-- Example: When "Adding a function to do X", explain why it is necessary to have a way to do X. -->

And seems that `atomic.Value` has better performance.
OLD:
```
Benchmark_Middleware_Logger-8             993579              1215 ns/op             281 B/op          3 allocs/op
Benchmark_Middleware_Logger-8            1000000              1184 ns/op             281 B/op          3 allocs/op
Benchmark_Middleware_Logger-8            1000000              1127 ns/op             281 B/op          3 allocs/op
Benchmark_Middleware_Logger-8            1000000              1120 ns/op             282 B/op          3 allocs/op
```

NEW:
```
Benchmark_Middleware_Logger-8            1000000              1151 ns/op             281 B/op          3 allocs/op
Benchmark_Middleware_Logger-8            1000000              1082 ns/op             281 B/op          3 allocs/op
Benchmark_Middleware_Logger-8            1000000              1069 ns/op             282 B/op          3 allocs/op
Benchmark_Middleware_Logger-8            1000000              1104 ns/op             282 B/op          3 allocs/op
```

**Commit formatting** 

Use emojis on commit messages so it provides an easy way of identifying the purpose or intention of a commit. Check out the emoji cheatsheet here: https://gitmoji.carloscuesta.me/